Update patch set 1
Patch Set 1:

The LTO file has 51 kb. IMO, give the option to choose when to download this data is something good (and have a wi-fi only option too).

I added this as hardware to allow supported devices to choose from where and to where the file is downloaded.

Let me know if still doesn't have sense for you, and then remove all this stuff and put it as an overlay.

Beside of that you +2'd the settings change :)

Patch-set: 1