Update patch set 1
Patch Set 1:

Does this mean that this will always be not supported? 

Or should somewhere there be a child class saying that it is supported as I can't seem to find the code that extends this class.

Patch-set: 1
Reviewer: Shuhao Wu <4229@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0